# docs/swagger.yaml
basePath: /api
definitions:
  handlers.authRequest:
    properties:
      email:
        example: example@vnest.org
        type: string
      password:
        example: superstrongpassword
        type: string
    type: object
...
  handlers.failedResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.successResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
    type: object
...
  handlers.userauthRequest:
    properties:
      email:
        example: example@vnest.org
        type: string
      password:
        example: superstrongpassword
        type: string
      position:
        example: founder
        type: string
    type: object
...
info:
  contact: {}
  description: This endpoint is for dev purposes
  title: VNEST Dashboard Swagger docs
  version: "1.0"
paths:
  /auth/user/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user by email and password. Uses a cache lookup
        before querying the database. Returns a JWT token on success.
      parameters:
      - description: User Login Input
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.authRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.successResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.failedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.failedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.failedResponse'
      summary: User Login
      tags:
      - auth
  /auth/user/signup:
    post:
      consumes:
      - application/json
      description: Registers a new user with an email and password. The user is assigned
        a default role of "user".
      parameters:
      - description: User Signup Input
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.userauthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.successResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.failedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.failedResponse'
      summary: User Signup
      tags:
      - auth
  /auth/vc/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user by email and password. Uses a cache lookup
        before querying the database. Returns a JWT token on success.
      parameters:
      - description: VC Login Input
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.authRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.successResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.failedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.failedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.failedResponse'
      summary: VC Login
      tags:
      - auth
  /auth/vc/signup:
    post:
      consumes:
      - application/json
      description: Registers a new VC user with an email and password. The user is
        assigned a default role of "vc".
      parameters:
      - description: VC Signup Input
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.authRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.successResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.failedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.failedResponse'
      summary: VC Signup
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" and your JWT token for authentication and authorization
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"